Move str's impl of iterations to new mod
